Ajustes Aula 07 de 20

+++ scr/mod_produto/Produto.py
@@ -1,8 +1,9 @@
 from pydantic import BaseModel
+from decimal import Decimal
 
 class Produto(BaseModel):
     id_produto: int = None
     nome: str
     descricao: str
     foto: bytes
-    valor_unitario: int
+    valor_unitario: Decimal
